refactor: extract interaction logic of SearchForm to new ViewModel
